requirements.txt
asgiref==3.8.1
confusable-homoglyphs==3.3.1
Django==5.0.3
django-bootstrap3==23.6
django-crispy-forms==2.1
django-registration==3.4
djangorestframework==3.15.1
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
